+++ lua/lush/transform/contrib/alacritty.lua
@@ -34,7 +34,7 @@
 --   dim_white = "#000000",
 -- }
 
-local helpers = require("lush.transformer.helpers")
+local helpers = require("lush.transform.helpers")
 local check_keys = {
   "fg", "bg",
   "cursor_fg", "cursor_bg",
