+++ spec/transform/viml_spec.moon
@@ -1,7 +1,5 @@
-describe "export.viml", ->
-  -- note, "export" is a protected keyword in moonscript
-  exporter = require("lush.exporter")
-  exp = exporter.export
+describe "transform.viml", ->
+  transform = require("lush.builder").transform
 
   setup ->
     parse = require('lush.parser')
@@ -20,13 +18,13 @@ describe "export.viml", ->
     package.loaded["theme"] = nil
 
   it "returns viml", ->
-    viml = require("lush.transformer.viml")
-    value = exp(require("theme"), viml)
+    viml = require("lush.transform.viml")
+    value = transform(require("theme"), viml)
     assert.is.table(value)
     assert.matches("highlight A guifg=NONE guibg=NONE guisp=NONE gui=italic blend=40", value[1])
 
   it "accepts options", ->
-    viml = require("lush.transformer.viml")
-    value = exp(require("theme"), {viml, {plugins: require("lush.compiler.plugin.vim_compatible")}})
+    viml = require("lush.transform.viml")
+    value = transform(require("theme"), {viml, {plugins: require("lush.compiler.plugin.vim_compatible")}})
     assert.is.table(value)
     assert.matches("highlight A guifg=NONE guibg=NONE guisp=NONE gui=italic blend=40", value[1])
